Replaced flux/redux with a simpler architecture
- We decided to stick to a simpler approach for version 1.0 during the
04/21 tech meeting.
