+++ package.json
@@ -7,8 +7,6 @@
   },
   "dependencies": {
     "react": "^0.14.8",
-    "react-native": "^0.24.1",
-    "react-redux": "^4.4.5",
-    "redux": "^3.5.1"
+    "react-native": "^0.24.1"
   }
 }
